pyproject.toml
[tool.poetry]
name = "carrousel"
version = "0.1.0"
description = ""
authors = ["Sasha K <a.kosak.mbx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.2"
pyyaml = "^6.0.1"
schema = "^0.7.5"
pathvalidate = "^3.2.0"
xscreensaver-config = "^1.0.7"
pillow = "^10.2.0"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
